+++ CHANGES.rst
@@ -3,6 +3,7 @@
 
 [//]: <> (Remove when new items added)
 - Add function to convert a jupyter notebook to html.
+- Add functions to strip a notebook or a folder of notebooks.
 
 
 [0.1.0](https://github.com/crim-ca/notebook-utils/tree/0.1.0) (30-09-2021)